Use a more human-friendly format for data files
Ordering the columns [code, price, label] makes it easier to define made-up datasets where the code is the same as the label. This comes at the expense of requiring an empty column when creating a data set that has labels but no prices. This feels like the right tradeoff, as it makes the simple case almost trivial and adds only a small amount of complexity to more complex cases.